+++ cypress/integration/console/shared/collaborators/add.spec.js
@@ -96,7 +96,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`user \`${userNotExist}\` not found`)
+        .findByText(userNotExist)
         .should('be.visible')
       cy.visit(
         `${Cypress.config('consoleRootPath')}/applications/${applicationId}/collaborators/add`,
@@ -124,7 +124,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`organization \`${orgNotExist}\` not found`)
+        .findByText(orgNotExist)
         .should('be.visible')
       cy.visit(
         `${Cypress.config('consoleRootPath')}/applications/${applicationId}/collaborators/add`,
@@ -169,7 +169,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`user \`${userNotExist}\` not found`)
+        .findByText(userNotExist)
         .should('be.visible')
       cy.visit(`${Cypress.config('consoleRootPath')}/gateways/${gatewayId}/collaborators/add`)
     })
@@ -195,7 +195,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`organization \`${orgNotExist}\` not found`)
+        .findByText(orgNotExist)
         .should('be.visible')
       cy.visit(`${Cypress.config('consoleRootPath')}/gateways/${gatewayId}/collaborators/add`)
     })
@@ -240,7 +240,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`user \`${userNotExist}\` not found`)
+        .findByText(userNotExist)
         .should('be.visible')
       cy.visit(`${Cypress.config('consoleRootPath')}/organizations/${testOrgId}/collaborators/add`)
     })
@@ -253,7 +253,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText('account of type `organization` can not collaborate on `organization`')
+        .findByText(/can not collaborate/i)
         .should('be.visible')
       cy.visit(`${Cypress.config('consoleRootPath')}/organizations/${testOrgId}/collaborators/add`)
     })
@@ -266,7 +266,7 @@ describe('Collaborators', () => {
 
       cy.findByTestId('error-notification')
         .should('be.visible')
-        .findByText(`organization \`${orgNotExist}\` not found`)
+        .findByText(orgNotExist)
         .should('be.visible')
       cy.visit(`${Cypress.config('consoleRootPath')}/organizations/${testOrgId}/collaborators/add`)
     })