+++ pkg/applicationserver/applicationserver_util_test.go
@@ -55,3 +55,34 @@ type noopEndDeviceFetcher struct{}
 func (noopEndDeviceFetcher) Get(ctx context.Context, ids ttnpb.EndDeviceIdentifiers, fieldMaskPaths ...string) (*ttnpb.EndDevice, error) {
 	return nil, errNotFound.New()
 }
+
+// MockLinkRegistry is a mock LinkRegistry used for testing.
+type MockLinkRegistry struct {
+	GetFunc   func(ctx context.Context, ids ttnpb.ApplicationIdentifiers, paths []string) (*ttnpb.ApplicationLink, error)
+	RangeFunc func(ctx context.Context, paths []string, f func(context.Context, ttnpb.ApplicationIdentifiers, *ttnpb.ApplicationLink) bool) error
+	SetFunc   func(ctx context.Context, ids ttnpb.ApplicationIdentifiers, paths []string, f func(*ttnpb.ApplicationLink) (*ttnpb.ApplicationLink, []string, error)) (*ttnpb.ApplicationLink, error)
+}
+
+// Get calls GetFunc if set and panics otherwise.
+func (m MockLinkRegistry) Get(ctx context.Context, ids ttnpb.ApplicationIdentifiers, paths []string) (*ttnpb.ApplicationLink, error) {
+	if m.GetFunc == nil {
+		panic("Get called, but not set")
+	}
+	return m.GetFunc(ctx, ids, paths)
+}
+
+// Range calls RangeFunc if set and panics otherwise.
+func (m MockLinkRegistry) Range(ctx context.Context, paths []string, f func(context.Context, ttnpb.ApplicationIdentifiers, *ttnpb.ApplicationLink) bool) error {
+	if m.RangeFunc == nil {
+		panic("Range called, but not set")
+	}
+	return m.RangeFunc(ctx, paths, f)
+}
+
+// Set calls SetFunc if set and panics otherwise.
+func (m MockLinkRegistry) Set(ctx context.Context, ids ttnpb.ApplicationIdentifiers, paths []string, f func(*ttnpb.ApplicationLink) (*ttnpb.ApplicationLink, []string, error)) (*ttnpb.ApplicationLink, error) {
+	if m.SetFunc == nil {
+		panic("Set called, bt not set")
+	}
+	return m.SetFunc(ctx, ids, paths, f)
+}