+++ pkg/applicationserver/grpc_deviceregistry_test.go
@@ -25,10 +25,12 @@ import (
 	"github.com/smartystreets/assertions"
 	. "go.thethings.network/lorawan-stack/v3/pkg/applicationserver"
 	"go.thethings.network/lorawan-stack/v3/pkg/auth/rights"
+	"go.thethings.network/lorawan-stack/v3/pkg/cluster"
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
 	componenttest "go.thethings.network/lorawan-stack/v3/pkg/component/test"
 	"go.thethings.network/lorawan-stack/v3/pkg/config"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
+	"go.thethings.network/lorawan-stack/v3/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/unique"
@@ -88,6 +90,8 @@ func TestDeviceRegistryGet(t *testing.T) {
 		t.Helper()
 		return assertions.New(t).So(errors.IsNotFound(err), should.BeTrue)
 	}
+
+	_, nsAddr := startMockNS(test.Context(), func(md rpcmetadata.MD) bool { return true })
 
 	for _, tc := range []struct {
 		Name            string
@@ -301,6 +305,12 @@ func TestDeviceRegistryGet(t *testing.T) {
 							Provider: "static",
 							Static:   registeredKEKs,
 						},
+						Cluster: cluster.Config{
+							NetworkServer: nsAddr,
+						},
+						GRPC: config.GRPC{
+							AllowInsecureForCredentials: true,
+						},
 					},
 				}),
 				&Config{
@@ -323,6 +333,9 @@ func TestDeviceRegistryGet(t *testing.T) {
 			})
 			componenttest.StartComponent(t, as.Component)
 			defer as.Close()
+
+			mustHavePeer(test.Context(), as.Component, ttnpb.ClusterRole_NETWORK_SERVER)
+			time.Sleep(2 * Timeout)
 
 			ctx := as.FillContext(test.Context())
 			req := deepcopy.Copy(tc.DeviceRequest).(*ttnpb.GetEndDeviceRequest)