+++ pkg/identityserver/api_key_utils.go
@@ -29,7 +29,8 @@ var apiKeyHashSettings auth.HashValidator = pbkdf2.PBKDF2{
 	SaltLength: 16,
 }
 
-func generateAPIKey(ctx context.Context, name string, rights ...ttnpb.Right) (key *ttnpb.APIKey, token string, err error) {
+// GenerateAPIKey generates a new API key with the given name for the set of rights
+func GenerateAPIKey(ctx context.Context, name string, rights ...ttnpb.Right) (key *ttnpb.APIKey, token string, err error) {
 	token, err = auth.APIKey.Generate(ctx, "")
 	if err != nil {
 		return nil, "", err