+++ CHANGELOG.md
@@ -27,6 +27,9 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Deprecated
 
+- Packet Broker mutual TLS authentication: use OAuth 2.0 client credentials instead; set `pba.authentication-mode` to `oauth2` and configure `pba.oauth2`.
+- Packet Broker forwarder blacklist setting `pba.home-network.blacklist-forwarder` has become ineffective.
+
 ### Removed
 
 - Application Server linking. The Network Server now pushes data to the cluster Application Server instead.
@@ -56,7 +59,6 @@ For details about compatibility between different releases, see the **Commitment
 ### Fixed
 
 - Removed misleading warning message for missing package data when setting up the storage integration package association.
-
 
 ## [3.10.4] - 2020-12-08